[jwxiong] implement maybe_repeat method

+++ src/include/restring.h
@@ -12,7 +12,7 @@ class REstring {
 public:
     explicit REstring(const std::string& pattern_): pattern(pattern_) {}
 
-    void remove_prefix(size_t count = 1);
+    void remove_prefix(int count = 1);
 
     size_t size() const;
